@fuwaneko 2016-07-22
Good point, various tests around the web seem to show that arguments object is always slower.